# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * ZOIKO Bible Explorer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DailyContent {
  verse: string;
  verseReference: string;
  challenge: string;
  memoryVerse: string;
  date: string;
}

export interface Category {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  questionCount?: number;
}

export type QuestionDifficulty = typeof QuestionDifficulty[keyof typeof QuestionDifficulty];


export const QuestionDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface Question {
  id: number;
  text: string;
  options: string[];
  correctAnswer: string;
  difficulty: QuestionDifficulty;
  categoryId: number;
  /** @nullable */
  explanation?: string | null;
  /** @nullable */
  book?: string | null;
  createdAt?: string;
}

export interface Player {
  name: string;
  correctAnswers: number;
  totalAnswers: number;
  sessionsPlayed: number;
  sessionsWon: number;
  sessionsHosted: number;
  streakCurrent: number;
  streakLongest: number;
  updatedAt: string;
  createdAt: string;
}

export interface PlayerInput {
  correctAnswers?: number;
  totalAnswers?: number;
  sessionsPlayed?: number;
  sessionsWon?: number;
  sessionsHosted?: number;
  streakCurrent?: number;
  streakLongest?: number;
}

export interface QuestionList {
  questions: Question[];
  total: number;
}

export type QuestionInputDifficulty = typeof QuestionInputDifficulty[keyof typeof QuestionInputDifficulty];


export const QuestionInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface QuestionInput {
  text: string;
  options: string[];
  correctAnswer: string;
  difficulty: QuestionInputDifficulty;
  categoryId: number;
  explanation?: string;
  book?: string;
}

export type GenerateQuestionsInputDifficulty = typeof GenerateQuestionsInputDifficulty[keyof typeof GenerateQuestionsInputDifficulty];


export const GenerateQuestionsInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
  mixed: 'mixed',
} as const;

export interface GenerateQuestionsInput {
  difficulty: GenerateQuestionsInputDifficulty;
  count: number;
  topic?: string;
}

export interface SaveQuestionsInput {
  questions: QuestionInput[];
}

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  waiting: 'waiting',
  active: 'active',
  completed: 'completed',
} as const;

export type SessionType = typeof SessionType[keyof typeof SessionType];


export const SessionType = {
  group: 'group',
  duel: 'duel',
} as const;

export type SessionDifficulty = typeof SessionDifficulty[keyof typeof SessionDifficulty];


export const SessionDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
  mixed: 'mixed',
} as const;

export interface Session {
  id: string;
  pin: string;
  hostCode: string;
  status: SessionStatus;
  type?: SessionType;
  difficulty: SessionDifficulty;
  /** @nullable */
  playStyle?: string | null;
  participants: string[];
  questions?: Question[];
  currentQuestionIndex: number;
  totalQuestions?: number;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
  /**
     * The authenticated caller's own display name for this session, if they have joined it. Never populated from client input — resolved server-side from the verified session_participants record — so the client can show "you" vs "opponent" without guessing.
     * @nullable
     */
  myDisplayName?: string | null;
}

export type SessionInputDifficulty = typeof SessionInputDifficulty[keyof typeof SessionInputDifficulty];


export const SessionInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
  mixed: 'mixed',
} as const;

export interface SessionInput {
  difficulty: SessionInputDifficulty;
  playStyle?: string;
  participants: string[];
  totalQuestions?: number;
}

export type SessionUpdateStatus = typeof SessionUpdateStatus[keyof typeof SessionUpdateStatus];


export const SessionUpdateStatus = {
  active: 'active',
  completed: 'completed',
} as const;

export interface SessionUpdate {
  status?: SessionUpdateStatus;
  currentQuestionIndex?: number;
}

export interface JoinSessionInput {
  pin: string;
  playerName: string;
}

export type DuelInputDifficulty = typeof DuelInputDifficulty[keyof typeof DuelInputDifficulty];


export const DuelInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
  mixed: 'mixed',
} as const;

export interface DuelInput {
  difficulty: DuelInputDifficulty;
  totalQuestions?: number;
  hostName: string;
}

export interface AnswerInput {
  questionId: number;
  answer: string;
}

export interface AnswerResult {
  correct: boolean;
  correctAnswer: string;
  score: number;
  totalScore: number;
  /** @nullable */
  explanation?: string | null;
}

export interface LeaderboardEntry {
  rank: number;
  playerName: string;
  score: number;
  correctAnswers: number;
  totalAnswers: number;
}

export interface Achievement {
  id: number;
  playerName: string;
  type: string;
  title: string;
  description: string;
  earnedAt: string;
}

export interface ActivityItem {
  type: string;
  description: string;
  createdAt: string;
}

export interface StatsOverview {
  totalQuestions: number;
  totalSessions: number;
  totalPlayers: number;
  questionsThisWeek: number;
  averageScore: number;
  topDifficulty?: string;
  recentActivity?: ActivityItem[];
}

export interface ErrorResponse {
  error: string;
}

export type ListQuestionsParams = {
difficulty?: ListQuestionsDifficulty;
categoryId?: number;
limit?: number;
offset?: number;
};

export type ListQuestionsDifficulty = typeof ListQuestionsDifficulty[keyof typeof ListQuestionsDifficulty];


export const ListQuestionsDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type ListRecentSessionsParams = {
limit?: number;
};

export type ListAchievementsParams = {
playerName?: string;
};