# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ZOIKO Bible Explorer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions with optional filters
      parameters:
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Questions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionList"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a new question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/generate:
    post:
      operationId: generateQuestions
      tags: [questions]
      summary: AI-generate Bible questions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateQuestionsInput"
      responses:
        "200":
          description: Generated questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
  /questions/generate/save:
    post:
      operationId: saveGeneratedQuestions
      tags: [questions]
      summary: Save AI-generated questions to the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveQuestionsInput"
      responses:
        "201":
          description: Saved questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get a question by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /players/{name}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get a player's aggregate stats, creating a record if it doesn't exist
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    put:
      operationId: upsertPlayer
      tags: [players]
      summary: Sync a player's aggregate stats (upserts to the max of existing vs incoming values)
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "200":
          description: Updated player stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
...
components:
  schemas:
...
    Question:
      type: object
      required: [id, text, options, correctAnswer, difficulty, categoryId]
      properties:
        id:
          type: integer
        text:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        categoryId:
          type: integer
        explanation:
          type: ["string", "null"]
        book:
          type: ["string", "null"]
        createdAt:
          type: string
    Player:
      type: object
      required: [name, correctAnswers, totalAnswers, sessionsPlayed, sessionsWon, sessionsHosted, streakCurrent, streakLongest, updatedAt, createdAt]
      properties:
        name:
          type: string
        correctAnswers:
          type: integer
        totalAnswers:
          type: integer
        sessionsPlayed:
          type: integer
        sessionsWon:
          type: integer
        sessionsHosted:
          type: integer
        streakCurrent:
          type: integer
        streakLongest:
          type: integer
        updatedAt:
          type: string
        createdAt:
          type: string
    PlayerInput:
      type: object
      properties:
        correctAnswers:
          type: integer
        totalAnswers:
          type: integer
        sessionsPlayed:
          type: integer
        sessionsWon:
          type: integer
        sessionsHosted:
          type: integer
        streakCurrent:
          type: integer
        streakLongest:
          type: integer
    QuestionList:
      type: object
      required: [questions, total]
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        total:
          type: integer
    QuestionInput:
      type: object
      required: [text, options, correctAnswer, difficulty, categoryId]
      properties:
        text:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        categoryId:
          type: integer
        explanation:
          type: string
        book:
          type: string
    GenerateQuestionsInput:
      type: object
      required: [difficulty, count]
      properties:
        difficulty:
          type: string
          enum: [easy, medium, hard, mixed]
        count:
          type: integer
          default: 5
        topic:
          type: string
    SaveQuestionsInput:
      type: object
      required: [questions]
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionInput"
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string