# lib/api-zod/src/generated/types/generateQuestionsInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * ZOIKO Bible Explorer API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateQuestionsInputDifficulty } from './generateQuestionsInputDifficulty';

export interface GenerateQuestionsInput {
  difficulty: GenerateQuestionsInputDifficulty;
  count: number;
  topic?: string;
}